+++ tap-snapshots/test/basic.js.test.cjs
@@ -334,11 +334,11 @@ exports[`test/basic.js TAP basic tests > makeRe @(.*|js) 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe @(a|a[(])b 1`] = `
-/^(?:(?=.)(?:(?!\\.)a|(?!\\.)a[(])b)$/
+/^(?:(?=.)(?:(?!\\.)a|(?!\\.)a\\()b)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe @(a|a[)])b 1`] = `
-/^(?:(?=.)(?:(?!\\.)a|(?!\\.)a[)])b)$/
+/^(?:(?=.)(?:(?!\\.)a|(?!\\.)a\\))b)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe @(js|.*) 1`] = `
@@ -406,7 +406,7 @@ exports[`test/basic.js TAP basic tests > makeRe [[:xdigit:]][[:xdigit:]][[:xdigi
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [[] 1`] = `
-/^(?:(?!\\.)(?=.)[\\[])$/
+/^(?:\\[)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [\\-\\]] 1`] = `
@@ -414,7 +414,7 @@ exports[`test/basic.js TAP basic tests > makeRe [\\-\\]] 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [\\\\] 1`] = `
-/^(?:(?!\\.)(?=.)[\\\\])$/
+/^(?:\\\\)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [\\b-a] 1`] = `
@@ -438,7 +438,7 @@ exports[`test/basic.js TAP basic tests > makeRe []-] 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe []] 1`] = `
-/^(?:(?!\\.)(?=.)[\\]])$/
+/^(?:\\])$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [^a-c]* 1`] = `
@@ -478,7 +478,7 @@ exports[`test/basic.js TAP basic tests > makeRe [abc-] 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [f-fz-a]* 1`] = `
-/^(?:(?!\\.)(?=.)[f][^/]*?)$/
+/^(?:(?=.)f[^/]*?)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [f-gz-a]* 1`] = `
@@ -486,7 +486,7 @@ exports[`test/basic.js TAP basic tests > makeRe [f-gz-a]* 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [fz-a]* 1`] = `
-/^(?:(?!\\.)(?=.)[f][^/]*?)$/
+/^(?:(?=.)f[^/]*?)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [ia]?[ck] 1`] = `
@@ -502,7 +502,7 @@ exports[`test/basic.js TAP basic tests > makeRe [z-a]* 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [z-af]* 1`] = `
-/^(?:(?!\\.)(?=.)[f][^/]*?)$/
+/^(?:(?=.)f[^/]*?)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe [z\\-a] 1`] = `
@@ -614,11 +614,11 @@ exports[`test/basic.js TAP basic tests > makeRe a[X-]b 1`] = `
 `
 
 exports[`test/basic.js TAP basic tests > makeRe a[\\b]c 1`] = `
-/^(?:(?=.)a[b]c)$/
+/^(?:abc)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe a[b]c 1`] = `
-/^(?:(?=.)a[b]c)$/
+/^(?:abc)$/
 `
 
 exports[`test/basic.js TAP basic tests > makeRe a\\*?/* 1`] = `